requirements.txt
numpy
pandas
scikit-learn
ipywidgets
ipykernel
googletrans
google-generativeai
groq
wxwidgets